--- backend/plugins/accounting_api/src/modules/inventories/graphql/resolvers/customResolvers/remainder.ts
@@ -1,11 +1,0 @@
-export default {
-  async category(remainder: any) {
-    if (!remainder.categoryId) {
-      return;
-    }
-    return {
-      __typename: 'ProductCategory',
-      _id: remainder.categoryId,
-    };
-  },
-};
